refactor(routes): extract asset routes
